requirements.txt
streamlit>=1.32
pandas
requests
python-dotenv
pillow
google-cloud-firestore
google-auth
google-api-core
